Move websocket test server to its own package
